@@ -1,6 +1,6 @@
 module github.com/fenglyu/adventofcode
 
-go 1.16
+go 1.18
 
 require (
 	github.com/deckarep/golang-set v1.7.1
